+++ database/migrations/2025_08_25_002809_create_item_sales_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->dateTime('tanggal_jual');
             $table->string('dijual_oleh')->nullable();
             $table->text('catatan')->nullable();
-            $table->double('total_harga')->default(0);
+            $table->unsignedBigInteger('total_harga')->default(0);
             $table->double('potongan_harga')->default(0);
             $table->string('status_pembayaran',20);
             $table->string('tipe_pembayaran',20);
